requirements.txt
streamlit==1.32.0
altair==4.2.2
langchain==0.2.14
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1
python-docx==1.1.2
transformers==4.44.2
torch==2.1.2
python-dotenv==1.0.1
pydantic==2.6.1
sentence-transformers==3.0.1
langchain-community==0.2.12
langchain-groq==0.1.9
docx2txt==0.8
pypdf==4.3.1

-e .